+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- next-header -->
+## [Unreleased] - ReleaseDate
 
-### Added
+## [0.7.0] - 2024-01-18
+
+### Changed
+- [breaking-change] Transitioned the `embedded-storage` interface to
+  use the new `embedded-hal` `DelayNs` trait instead of the old `CountDown`
+- [breaking-change] Transitioned `embedded-hal` to version 1.0
+- Updated `embedded-storage` to version 0.3.1
 - Add support for 24AA02xE48/E64 devices.
 
 ## [0.6.1] - 2023-12-27
@@ -83,7 +90,9 @@ This is the initial release to crates.io of the feature-complete driver. There
 may be some API changes in the future, in case I decide that something can be
 further improved. All changes will be documented in this CHANGELOG.
 
-[Unreleased]: https://github.com/eldruin/eeprom24x-rs/compare/v0.6.1...HEAD
+<!-- next-url -->
+[Unreleased]: https://github.com/eldruin/eeprom24x-rs/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/eldruin/eeprom24x-rs/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/eldruin/eeprom24x-rs/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/eldruin/eeprom24x-rs/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/eldruin/eeprom24x-rs/compare/v0.4.0...v0.5.0